requirements.txt
pyautogui==0.9.53
pillow==9.5.0
requests==2.31.0
pydantic==2.4.2
opencv-python==4.8.0.74
google-generativeai>=0.3.0
python-dotenv>=1.0.0 